requirements.txt
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
pyrebase4==4.5.0

